@@ -1,12 +1,20 @@
-import Vue from 'vue'
-import App from './App.vue'
-import router from './router'
-import store from './store'
+import { Container } from 'vue-typedi'
+import StoreModules from '@/store/CreateStoreModules'
+import WebSocket from '@/ws/WebSocket'
+import createApp from '@/app'
+import tokens from '@/tokens'
 
-Vue.config.productionTip = false
+const store: StoreModules = Container.get(tokens.STORE) as StoreModules
+const websocket: WebSocket = Container.get(tokens.WEBSOCKET) as WebSocket
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+store.theme.setDarkTheme(store.theme.getThemeIsDark)
+store.auth.setHttpAccessToken()
+
+if (!store.auth.tokensUpdatedLongTime()) {
+  store.auth.setAuthorizationUser(true)
+}
+
+createApp()
+
+// eslint --fix --ext .js, .ts, .vue src
+// eslint --fix --ext .js,.vue,.ts src
